Pull request

+++ xchange-bitstamp/src/main/java/org/knowm/xchange/bitstamp/dto/marketdata/BitstampTransaction.java
@@ -7,7 +7,7 @@ import java.math.BigDecimal;
 public class BitstampTransaction {
 
   private final long date;
-  private final int tid;
+  private final long tid;
   private final BigDecimal price;
   private final BigDecimal amount;
   private final int type;
@@ -22,7 +22,7 @@ public class BitstampTransaction {
    */
   public BitstampTransaction(
       @JsonProperty("date") long date,
-      @JsonProperty("tid") int tid,
+      @JsonProperty("tid") long tid,
       @JsonProperty("price") BigDecimal price,
       @JsonProperty("amount") BigDecimal amount,
       @JsonProperty("type") int type) {
@@ -34,7 +34,7 @@ public class BitstampTransaction {
     this.type = type;
   }
 
-  public int getTid() {
+  public long getTid() {
 
     return tid;
   }
